+++ tests/testthat/test-prioritize.R
@@ -217,10 +217,10 @@ test_that("passes if score=1 & values by id_loan are duplicated for distinct
 test_that("with 0-row input returns 0-row input", {
   lbk <- fake_lbk()
   ald <- fake_ald(name_company = "won't match")
-  cero_row <- suppressWarnings(match_name(lbk, ald))
+  zero_row <- suppressWarnings(match_name(lbk, ald))
 
-  has_cero_row <- identical(nrow(cero_row), 0L)
-  stopifnot(has_cero_row)
+  has_zero_row <- identical(nrow(zero_row), 0L)
+  stopifnot(has_zero_row)
 
-  expect_no_error(prioritize(cero_row))
+  expect_no_error(prioritize(zero_row))
 })